backend/requirements.txt
# Core
Flask>=3.1.0
flask-cors>=6.0.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0

# Database
alembic>=1.13.0
Flask-Migrate>=4.0.0

# Storage (S3/MinIO)
boto3>=1.34.0

# Authentication – Microsoft Entra ID (Azure AD) JWT validation
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
requests>=2.31.0
